% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GESTIA.R
\name{adjustImpactScore}
\alias{adjustImpactScore}
\title{Adjust the impact score}
\usage{
adjustImpactScore(impact.score, meanImpScore)
}
\arguments{
\item{impact.score}{The raw impact score}

\item{meanImpScore}{the mean of the permutated impact score}
}
\value{
GESTIA score after adjustment
}
\description{
This function first check whether both of the impact score and permutated scores are 0, then adjust the normalized GESTIA score
}
