% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GESTIA.R
\name{calculateMa}
\alias{calculateMa}
\title{Calculate matrix of pairwise pathway impact of row over column}
\usage{
calculateMa(list.gs, ig.net, imCut = 0, penal.isolated = 0.1,
  plot = FALSE, n.permu = 30)
}
\arguments{
\item{list.gs}{The list of genes of each pathway/gene set.}

\item{ig.net}{The global network}

\item{imCut}{The cutoff of impact scores. Default 0}

\item{penal.isolated}{The penalization factor for isolated genes. Default 0.1}

\item{plot}{Whether to plot the splitted network. Default FALSE}

\item{n.permu}{The number of permutations for calculation of empirical null distribution}
}
\value{
A list of two matrices, the normalized GESTIA scores and the p-values
}
\description{
A wrapper to calculate the GESTIA matrix of a list of pathways. The core input is the list of pathways
}
