% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GESTIA.R
\name{core_rImp}
\alias{core_rImp}
\title{Calculation of the raw impact score}
\usage{
core_rImp(ig.ab, degree.a, degree.b, rp.a.d, rp.b.d, rp.a.u, rp.b.u,
  penal.isolated = 0.1, cutImpactSub = 0, exportData = F)
}
\arguments{
\item{ig.ab}{The network of the two pathways}

\item{degree.a}{The degree of each genes in pathway A}

\item{degree.b}{The degree of each genes in pathway B}

\item{rp.a.d}{The proportion of downstream genes for each gene in pathway A}

\item{rp.b.d}{The proportion of downstream genes for each gene in pathway B}

\item{rp.a.u}{The proportion of upstream genes for each gene in pathway A}

\item{rp.b.u}{The proportion of upstream genes for each gene in pathway B}

\item{penal.isolated}{The penalization factor for isolated genes. Default 0.1}

\item{cutImpactSub}{The cutoff for impact score. Default 0}

\item{exportData}{Whether to output details of the calculation. Default F}
}
\value{
A list containing: Raw GESTIA score, the impact score of A on B, and the impact score of B on A
}
\description{
Calculation of the raw impact score. The core of the GESTIA
}
