\name{GESTIA-package}
\alias{GESTIA-package}
\alias{GESTIA}
\docType{package}
\title{
GEne Set Impact Analysis to measure upstream/downstream correlations of pathways
}
\description{
\packageDescription{GESTIA}
}
\details{
The main function of this package is gestia, which calculates a score to measure the upstream/downstream correlations of pathways.
}
\author{
\packageAuthor{GESTIA}

Maintainer: \packageMaintainer{GESTIA}
}
\references{
https://github.com/yanshen2953/GESTIA
}
\keyword{ package }
\examples{
#See vignettes for full calculation of GESTIA
}
