% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GESTIA.R
\name{gestia}
\alias{gestia}
\title{GESTIA calculation}
\usage{
gestia(g.tmp, d1.symbol, d2.symbol, penal.isolated = 0.1,
  cutImpact = 0, plot = FALSE, n.permu = 30, minGeneNum = 5,
  exportData = F)
}
\arguments{
\item{g.tmp}{Global network constructed from KEGG}

\item{d1.symbol}{Gene names of pathway A}

\item{d2.symbol}{Gene names of pathway B}

\item{penal.isolated}{Penalization factor for isolated genes. Default 0.1}

\item{cutImpact}{Cutoff for impact score. Scores less than this cutoff will be set to 0. Default 0.}

\item{plot}{Whether to plot the splitted network. Default FALSE}

\item{n.permu}{Number of permutations for calculating the empirical null distribution}

\item{minGeneNum}{Cutoff for number of genes in pathways after id convertion}

\item{exportData}{Whether to export detail of calculation result. Default FALSE.}
}
\value{
List of results including: p-value, Normalized score,ScoreA before normalization,ScoreB before normalization,Mean permutation A,Mean permutation B
}
\description{
The main function of GESTIA.
}
