% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GESTIA.R
\name{giop}
\alias{giop}
\title{Calculate each gene's impact(from pathway A) over pathwayB: gene impact over pathway->giop}
\usage{
giop(ig.tmp, ids.1.tmp, ids.2.tmp, rp.2.tmp)
}
\arguments{
\item{ig.tmp}{The combined network of the two pathways}

\item{ids.1.tmp}{The gene symbols of pathway A}

\item{ids.2.tmp}{The gene symbols of pathway B}

\item{rp.2.tmp}{The relative position of genes in pathway B, i.e. the proportion of downstream genes for each gene in pathway B}
}
\value{
A vector containing the impact of each gene in pathway A
}
\description{
Calculate the impact of each gene in pathway A on pathway B
}
