% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GESTIA.R
\name{log2Transform}
\alias{log2Transform}
\title{Log transformation of the matrix}
\usage{
log2Transform(ma, cut = 0.1)
}
\arguments{
\item{ma}{The matrix to be transformed}

\item{cut}{The cutoff for each value in the matrix after transformation}
}
\value{
The matrix after transformation
}
\description{
Log transformation of the GESTIA matrix, with sign maintained.
}
