% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GESTIA.R
\name{mergePw}
\alias{mergePw}
\title{Merge two pathways' network}
\usage{
mergePw(ig.all, list.pw)
}
\arguments{
\item{ig.all}{The global network}

\item{list.pw}{A list of the pathways to merge. For example, list.pw[[1]] contains gene symbols of pathway A's genes.}
}
\value{
An igraph object of the merged network
}
\description{
Merge two pathways' network by deleting unrelated genes. This strategy can maintain the inter-pathway interactions.
}
