% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GESTIA.R
\name{prepareMaPlot}
\alias{prepareMaPlot}
\title{Convert GESTIA score matrix to igraph object}
\usage{
prepareMaPlot(ls.ma, pCut = 0.05, scoreCut = 0)
}
\arguments{
\item{ls.ma}{The list of two matrices, first is the GESTIA scores, second is the p-values}

\item{pCut}{The cutoff for p-value}

\item{scoreCut}{The cutoff for GESTIA score}
}
\value{
An igraph object of the GESTIA score matrix
}
\description{
Filter and then convert GESTIA score matrix to an igraph object, using the matrix as adjacency matrix
}
