% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GESTIA.R
\name{pValueCal}
\alias{pValueCal}
\title{Calculate p value}
\usage{
pValueCal(score.a, score.b, a.permu, b.permu)
}
\arguments{
\item{score.a}{The raw impact score of pathway A on B}

\item{score.b}{The raw impact score of pathway B on A}

\item{a.permu}{The permutated impact scores of pathway A on B}

\item{b.permu}{The permutated impact scores of pathway B on A}
}
\value{
p-value
}
\description{
Calculate the p-value based on the raw and permutated impact scores.
}
