% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GESTIA.R
\name{relativeImpact2}
\alias{relativeImpact2}
\title{Calculate impact of pathway A over pathway B}
\usage{
relativeImpact2(ig.all, degree.a, degree.b, rp.a.d, rp.b.d, rp.a.u, rp.b.u,
  penal.isolated = 0.1, cutImpactSub = 0, plot = FALSE,
  n.permu = 30, minGeneNum = 5, exportData = F)
}
\arguments{
\item{ig.all}{Global network constructed from KEGG}

\item{degree.a}{The degree of genes in pathway A}

\item{degree.b}{The degree of genes in pathway B}

\item{rp.a.d}{The proportion of downstream genes of the genes in pathway A}

\item{rp.b.d}{The proportion of downstream genes of the genes in pathway B}

\item{rp.a.u}{The proportion of upstream genes of the genes in pathway A}

\item{rp.b.u}{The proportion of upstream genes of the genes in pathway B}

\item{penal.isolated}{The penalization factor for the isolated genes. Default is 0.1}

\item{cutImpactSub}{The cutoff for impact score. Impact score smaller than this cutoff will be set to 0. Default 0.}

\item{plot}{Whether to plot the splitted network of the two pathways. TRUE or FALSE. Default FALSE.}

\item{n.permu}{Number of permutations for calculating the empirical null distribution. Default 30.}

\item{minGeneNum}{The cutoff for number of genes in pathways. Pathways with less genes (after id conversion) than this cutoff will be disgarded.}

\item{exportData}{Whether or not to export more details in the calculation. Default FALSE.}
}
\value{
List of results including: p-value, Normalized score,ScoreA before normalization,ScoreB before normalization,Mean permutation A,Mean permutation B
}
\description{
This function calls the core of the calculation function, and calculate the raw and permutated impact scores, then calculate p-value and GESTIA scores.
}
