% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GESTIA.R
\name{relativePosition}
\alias{relativePosition}
\title{calculate the relative positions of genes in a pathway}
\usage{
relativePosition(id.vertex, tmp.g)
}
\arguments{
\item{id.vertex}{A vector containing the names of the genes}

\item{tmp.g}{An igraph object which was constructed from the global network by deleting the unrelated genes of the pathway}
}
\value{
A dataframe consisted of two variables, the first is the proportion of downstream genes, the second is the proportion of upstream genes.
}
\description{
Calculate the ratio of upstream/downstream genes in a pathway. For scattered pathway/gene set, the ratio is weighted by the proportion of the sub-graph.
}
