% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GESTIA.R
\name{set_splitPwColor}
\alias{set_splitPwColor}
\title{Set color for the splitted pathways for better visualization}
\usage{
set_splitPwColor(ig.split, gss1, gss2, verticesColor = TRUE)
}
\arguments{
\item{ig.split}{The splitted network, an igraph object}

\item{gss1}{The gene symbols of pathway A's genes}

\item{gss2}{The gene symbols of pathway B's genes}

\item{verticesColor}{Whether to paint the vertices with color. Default TRUE}
}
\value{
An igraph object with colors
}
\description{
Set the color for the vertices and edges in the splitted network.
}
