% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GESTIA.R
\name{split_intersected_pathways}
\alias{split_intersected_pathways}
\title{Split two intersected pathways}
\usage{
split_intersected_pathways(ig.all, genes1, genes2)
}
\arguments{
\item{ig.all}{The global network}

\item{genes1}{The gene symbols of pathway A's genes}

\item{genes2}{The gene symbols of pathway B's genes}
}
\value{
A list containing: the igraph object of the splitted network, the new genes1, the vector genes2
}
\description{
Split two intersected pathways by duplicating the shared genes. Genes1 will be modified to contain fake vertices with "fa_"
}
