% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_GESTIA.R
\name{visual2pw}
\alias{visual2pw}
\title{Visualize two pathways' components}
\usage{
visual2pw(ig.all, genesa, genesb, colorset = c("yellowgreen", "tomato",
  "gold"), isolated = FALSE, ly = c("kk", "fr", "random", "grid"))
}
\arguments{
\item{ig.all}{The global network}

\item{genesa}{The gene symbols of pathway A}

\item{genesb}{The gene symbols of pathway B}

\item{colorset}{The color set for the shared genes, genes of pathway A, and the genes of pathway B. Default yellowgreen, tomato, gold}

\item{isolated}{Whether to plot the isolated genes. Default FALSE}

\item{ly}{Layout of the igraph. Available options: kk, fr, random, grid}
}
\value{
Plot directly.
}
\description{
Visualization of two pathways.
}
